requirements.txt
Flask_Bcrypt==0.7.1
Flask_Login==0.5.0
WTForms==2.3.1
Flask_SQLAlchemy==2.4.4
itsdangerous==1.1.0
Flask==1.1.1
Flask_Mail==0.9.1
Flask_WTF==0.14.3
Pillow==7.2.0
secrets==1.0.2
